+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Time Block changelog
 
-## 0.0.17 -- 2022-09-16 23:29:49
+## 0.0.17 -- 2022-09-17 01:17:28
 
 - Fix column layout issue - had delete code that was needed
 - Prevent excess renders when showing the details popup
 - Remove the search context and push into the main state
 - Remove the color scale context
+- Move hotkey code into the store to prevent excess renders
 
 ## 0.0.16 -- 2022-09-13 23:43:52